chore(asset-import-scripts): refactor asset import script files
- remove redundant scripts files
- adds more validation is_valid_file validator
- moves script files and csv's to utils folder
