+++ utils/asset/post_scripts.py
@@ -4,12 +4,14 @@ import csv
 from tqdm import tqdm
 import django
 
-from .helpers import display_inserted, display_skipped
+from utils.helpers import display_inserted, display_skipped
 
 project_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(project_dir)
+
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings")
 django.setup()
+
 from core.models.asset import (
     AssetType, AssetMake, Asset, AssetModelNumber, AssetCategory,
     AssetSubCategory,
